Modified Test APIs
