+++ tests/test_APIs.py
@@ -1,24 +1,13 @@
 import pytest
 import allure
-from apis import APIs
 from utils.logger import LogGen
 from tests.helpers import Helpers
 from utils.string_utils import generate_random_email
 
 logger = LogGen.loggen()
 
-@pytest.fixture(scope="module")
-def apis(auth_token):
-    logger.info("Creating APIs instance")
-    return APIs(auth_token)
 
-@pytest.fixture(scope="module")
-def user_id(apis):
-    logger.info("Fixture Auth Token & User ID")
-    response = Helpers.post_user(apis)
-    user_id = response[0]
-    return user_id
-
+@pytest.mark.high
 @pytest.mark.api
 @allure.severity(allure.severity_level.CRITICAL)
 def test_get_users(apis):
@@ -26,6 +15,7 @@ def test_get_users(apis):
     json_str = Helpers.get_users(apis)
     logger.info(f"json GET response body: {json_str}")
     logger.info("Get user is done.....")
+
 
 @pytest.mark.high
 @pytest.mark.api
@@ -43,6 +33,7 @@ def test_post_user(apis, user_data):
     logger.info(f"json POST response body: {json_str}")
     logger.info("Post user is done.....")
 
+
 @pytest.mark.high
 @pytest.mark.api
 @pytest.mark.parametrize(
@@ -58,6 +49,7 @@ def test_put_user(apis, user_id, user_data):
     logger.info(f"json PUT response body: {json_str}")
     logger.info("Update user is done.....")
 
+
 @pytest.mark.high
 @pytest.mark.api
 @allure.severity(allure.severity_level.NORMAL)
@@ -67,6 +59,7 @@ def test_delete_user(apis, user_id):
     logger.info("Delete user is done.....")
     logger.info("End of API Testing")
 
+
 @pytest.mark.low
 @pytest.mark.file
 @allure.severity(allure.severity_level.MINOR)
@@ -74,6 +67,7 @@ def test_upload_file(apis):
     logger.info("Uploading the file.......")
     Helpers.upload_file(apis)
     logger.info("Successfully uploaded the file")
+
 
 @pytest.mark.low
 @pytest.mark.file
